buildメソッド等の整理・ファイル名の変更
