+++ lib/mercari/components/item_parts/item_button.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+import 'package:flutter_tutorial/constants.dart';
+
+class ItemButton extends StatelessWidget {
+  const ItemButton({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+      style: ElevatedButton.styleFrom(
+        primary: freeMarketRedAccentColor,
+        onPrimary: Colors.white,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(4),
+        ),
+      ),
+      onPressed: () {},
+      child: const Text(
+        '出品する',
+        style: TextStyle(
+          fontWeight: FontWeight.w600,
+          fontSize: 12,
+        ),
+      ),
+    );
+  }
+}
